Move screenlock to main.tsx
