+++ src/pages/Layout.tsx
@@ -18,23 +18,6 @@ const Layout = () => {
         }
     }, []);
 
-    useEffect(() => {
-        async function lockScreen() {
-            try {
-                const screen: any = window.screen;
-
-                screen.lockOrientationUniversal = screen.lockOrientation || screen.mozLockOrientation || screen.msLockOrientation;
-
-                if ('lock' in screen.orientation) await (screen.orientation as any).lock('portrait-primary');
-                else if (screen.lockOrientationUniversal) screen.lockOrientationUniversal('portrait-primary');
-                else console.error("Orientation lock not supported");
-            } catch (e) {
-                console.error("Error trying to lock orientation", e);
-            }
-        }
-        lockScreen();
-    }, []);
-
     return (
         <div className="w-full h-dvh relative flex flex-col overflow-hidden">
             <div className="w-full h-full overflow-y-auto">
